+++ FireStep/Display.h
@@ -18,6 +18,7 @@ enum DisplayStatus {
     DISPLAY_BUSY = 30, // FireStep is busy but not moving
     DISPLAY_BUSY_MOVING = 31, // FireStep is busy and moving
     DISPLAY_BUSY_CALIBRATING = 32, // FireStep is busy and calibrating
+	DISPLAY_EEPROM = 33, // Accessing EEPROM
 };
 
 typedef class Display {
@@ -37,6 +38,9 @@ public:
     inline DisplayStatus getStatus() {
         return (DisplayStatus) status;
     }
+    virtual DisplayStatus getStatus() {
+		return this->status;
+	}
     virtual void setStatus(DisplayStatus status = DISPLAY_WAIT_IDLE) {
         this->status = status;
         show();